+++ app/assets/javascripts/imgupload.js
@@ -9,8 +9,6 @@ ImageUploadHandler.prototype = Object.create(UploadHandler.prototype);
 
 ImageUploadHandler.prototype.createAssetRecord = function(postdata) {
     postdata[pss_asset_type]['size'] = $('input[name="image[size]"]:checked').val();
-    postdata[pss_asset_type]['width'] = $('input[name="image[width]"]').val();
-    postdata[pss_asset_type]['height'] = $('input[name="image[height]"]').val();
     postdata[pss_asset_type]['alt_text'] = $('input[name="image[alt_text]"]').val();
 
     this.postAssetRecord(postdata);