+++ app/helpers/application_helper.rb
@@ -1,9 +1,5 @@
 module ApplicationHelper
 
-  def source_name(source)
-    source.name.present? ? source.name : source.aggregation
-  end
-
   ##
   # Returns frontend path with correctly joining '/'
   # @param path String (with or without leading slash)
@@ -56,18 +52,6 @@ module ApplicationHelper
       end
     else
       image_name
-    end
-  end
-
-  ##
-  # Get all the authors for an object with related authors
-  # @param Guide or SourceSet
-  # @return Array
-  def authors(authored)
-    return unless authored.present?
-    authored.authors.map do |author|
-      author.affiliation.present? ? author.name + ', ' + author.affiliation :
-        author.name
     end
   end
 end