+++ spec/javascripts/image_upload_handler_spec.js
@@ -22,35 +22,12 @@ describe("ImageUploadHandler", function() {
                     file_name: 'a.jpg',
                     source_ids: '4',
                     size: 'small',
-                    width: undefined,
-                    height: undefined,
                     alt_text: undefined
                 }
             };
             var html =
                 '<input name="image[size]" type="radio" value="large">' +
                 '<input name="image[size]" type="radio" value="small" checked>';
-            appendSetFixtures(html);
-            iuh.createAssetRecord(input_data);
-            expect(iuh.postAssetRecord).toHaveBeenCalledWith(good_data);
-        }
-    );
-
-    it("Generates AJAX POST data with width and height",
-        function() {
-            var good_data = {
-                image: {
-                    file_name: 'a.jpg',
-                    source_ids: '4',
-                    size: undefined,
-                    width: '600',
-                    height: '400',
-                    alt_text: undefined
-                }
-            };
-            var html =
-               '<input name="image[width]" type="text" value="600" />' +
-               '<input name="image[height]" type="text" value="400" />';
             appendSetFixtures(html);
             iuh.createAssetRecord(input_data);
             expect(iuh.postAssetRecord).toHaveBeenCalledWith(good_data);
@@ -64,8 +41,6 @@ describe("ImageUploadHandler", function() {
                     file_name: 'a.jpg',
                     source_ids: '4',
                     size: undefined,
-                    width: undefined,
-                    height: undefined,
                     alt_text: 'x'
                 }
             };