+++ spec/views/audios/show.html.erb_spec.rb
@@ -8,6 +8,8 @@ describe 'audios/show.html.erb', type: :view do
     assign(:audio, audio)
   end
 
+  it_behaves_like 'renderable view'
+
   it 'renders the audio' do
     render
     expect(rendered).to include(audio.file_base)