+++ spec/views/images/show.html.erb_spec.rb
@@ -8,6 +8,8 @@ describe 'images/show.html.erb', type: :view do
     assign(:image, image)
   end
 
+  it_behaves_like 'renderable view'
+
   it 'renders the image' do
     render
     expect(rendered).to include(image.file_name)