+++ spec/views/sources/show.html.erb_spec.rb
@@ -9,6 +9,8 @@ describe 'sources/show.html.erb', type: :view do
     assign(:source_set, source.source_set)
   end
 
+  it_behaves_like 'renderable view'
+
   it 'renders the source' do
     render
     expect(rendered).to include(source.aggregation)