+++ spec/views/tags/edit.html.erb_spec.rb
@@ -1,0 +1,6 @@
+require 'rails_helper'
+
+describe 'tags/edit.html.erb', type: :view do
+  before { assign(:tag, create(:tag_factory)) }
+  it_behaves_like 'renderable view'
+end